// packages/backend/convex/_generated/api.d.ts
export declare const components: {
  betterAuth: {
    adapter: {
      create: FunctionReference<
        "mutation",
        "internal",
        {
          input:
            | {
                data: {
                  banExpires?: null | number;
                  banReason?: null | string;
                  banned?: null | boolean;
                  createdAt: number;
                  email: string;
                  emailVerified: boolean;
                  image?: null | string;
                  name: string;
                  role?: null | string;
                  tier?: null | string;
                  updatedAt: number;
                  userId?: null | string;
                };
// ...
              };
          onCreateHandle?: string;
          select?: Array<string>;
        },
// ...
    };
    organisation: {
      getUsersOwnedOrganisation: FunctionReference<
        "query",
        "internal",
        { userId: string },
        Array<{
          createdAt: number;
          organizationId: string;
          role: string;
          userId: string;
        }>
      >;
    };
  };
  todo: {};
};